Separate some test files
